+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the "pivot-lang-support" extension will be documented in 
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.0.27]
+
+- Add Pivot Lang jupyter notebook support
+  - works on `*.iplnb`
+
 ## [0.0.26]
 
 - even better textmate syntax highlighting